fix: regdid check status on article/subarticle new/delete

+++ crud/regdid/views.py
@@ -334,6 +334,9 @@ def regdid_articles_new(request, regdid_id, article_num):
     didatticacdsarticoliregolamentoform = DidatticaCdsArticoliRegolamentoForm(data=request.POST if request.POST else None)
     testata = get_object_or_404(DidatticaCdsArticoliRegolamentoTestata, cds_id=regdid.cds, aa=regdid.aa_reg_did)
     
+    if testata.id_didattica_articoli_regolamento_status.status_cod in ['2', '3']:
+        return custom_message(request, _("Permission denied"))
+    
     # permissions    
     user_permissions_and_offices = testata.get_user_permissions_and_offices(request.user)
     if not user_permissions_and_offices['permissions']['access']:
@@ -410,6 +413,10 @@ def regdid_articles_delete(request, regdid_id, article_id):
     titolo_articolo = articolo.id_didattica_articoli_regolamento_struttura.titolo_it
     numero_sotto_art = DidatticaCdsSubArticoliRegolamento.objects.filter(id_didattica_cds_articoli_regolamento=article_id).count()
     testata = get_object_or_404(DidatticaCdsArticoliRegolamentoTestata, cds_id=regdid.cds, aa=regdid.aa_reg_did)
+    
+    if testata.id_didattica_articoli_regolamento_status.status_cod in ['2', '3']:
+        return custom_message(request, _("Permission denied"))
+    
     articolo.delete()
     
     log_action(user=request.user,
@@ -519,6 +526,10 @@ def regdid_sub_articles_new(request, regdid_id, article_id):
     didatticacdssubarticoliregolamentoform = DidatticaCdsSubArticoliRegolamentoForm(data=request.POST if request.POST else None)
     strutt_articolo = articolo.id_didattica_articoli_regolamento_struttura
     testata = get_object_or_404(DidatticaCdsArticoliRegolamentoTestata, cds_id=regdid.cds, aa=regdid.aa_reg_did)
+    
+    if testata.id_didattica_articoli_regolamento_status.status_cod in ['2', '3']:
+        return custom_message(request, _("Permission denied"))
+    
     note_revisione = articolo.note
     
     # Nav-bar items
@@ -587,6 +598,10 @@ def regdid_sub_articles_delete(request, regdid_id, article_id, sub_article_id):
     titolo_sotto_articolo = sotto_articolo.titolo_it
     ordine_sotto_articolo = sotto_articolo.ordine
     testata = get_object_or_404(DidatticaCdsArticoliRegolamentoTestata, cds_id=regdid.cds, aa=regdid.aa_reg_did)
+    
+    if testata.id_didattica_articoli_regolamento_status.status_cod in ['2', '3']:
+        return custom_message(request, _("Permission denied"))
+    
     sotto_articolo.delete()
     
     log_action(user=request.user,
